removendo a opção de alterar o nome do elemento

+++ src/tabelaSimbolos/Prime.java
@@ -5,8 +5,8 @@ import java.math.BigInteger;
 public class Prime {
 
 	// Function to get nextPrimeNumber
-	static long nextPrime(Integer n) {
+	static int nextPrime(Integer n) {
 		BigInteger b = new BigInteger(String.valueOf(n));
-		return Long.parseLong(b.nextProbablePrime().toString());
+		return Integer.parseInt(b.nextProbablePrime().toString());
 	}
 }
